fix(migrate): verify Hermes ledger evidence
